requirements.txt
Flask>=2.3
Flask-SQLAlchemy
Werkzeug
gunicorn
requests
apscheduler
python-dotenv>=1.0.0
cryptography>=41.0.0
pybitlaunch>=1.0.0
psycopg2-binary>=2.9.0
flask-migrate>=4.0.0
flask-cors>=4.0.0 